requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Admin
Flask-WTF
Flask-Migrate
email-validator
python-dotenv
Flask-Bcrypt
Flask-Mail
Flask-Caching
redis
psutil
